1.8.04 Add Difficulty Select Screen
I had to look at the solution to find out that I needed to extend
InputAdapter then implement Screen. I would not have found this out
without hour in the libgdx docs. In the execercise the entire class was
given so I don't feel too bad for having to look.